Fix IMAP response ) parsing
I was a bit confounded that the responses back from the UID command
have more responses than I gave them UIDs.

It turns out that there are either N*2 or N*2+1 responses from those
commands.  Each UID gets a response in the form of a tuple, but each
of _those_ is followed by a plain 'str' object filled with ')'.

But, sometimes, the whole sequence is followed by _another_ string.
That string seems to be for the flags that all the messages share,
like: "FLAGS (\Seen))"

But, I screwed it up, and wasn't looking for these responses, and
went out of range on the uids array.  I should _probably_ be parsing
these and looking for all the matching '(' ')'  pairs in the
responses.  But, for now, just ignore all the string responses, and
try to parse the tuples.