Add some reliability
This wraps quite a few of the IMAP commands.  It then notices
when exceptions (such as network errors) or transient conditions
(like append errors when the account is not full).  It will tear
down the IMAP connection, and open a new one in its place,
hopefully without any of the rest of the software noticing.

This should let these filesystems stay connected for much longer
and also better tolerate some of the errors that GMail is starting
to throw these days.